# Conflicts:
#	Assets/Script/GamePlay/ModifyTrigger/TriggerData/ModifyTriggerData.cs
#	Assets/Script/GamePlay/Ship/Controller/JobManager.cs
#	Assets/Script/Manager/AIManager.cs

+++ Assets/Script/Config/GameMiscConfig.cs
@@ -105,7 +105,7 @@ public enum ShipClassType
 public class BattleMiscRefreshConfig
 {
     public int LevelUpRefreshCostBase;
-    public int LevelUpCostWaveMultiple;
+    public float LevelUpCostWaveMultiple;
     public int LevelUpRefreshCountMultiple;
 
     public float Harbor_Teleport_RandomRangeMin;
